separated the shooter test command to shooter/feeder/pivot, changed trigger subsystem name to feeder
