+++ src/main/java/frc/robot/RobotContainer.java
@@ -7,8 +7,15 @@ package frc.robot;
 import frc.robot.Constants.OperatorConstants;
 
 import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
+import frc.robot.subsystems.Pivot;
+import frc.robot.subsystems.*;
+import frc.robot.commands.PivotTestCommand;
+import frc.robot.commands.ShooterTestCommand;
+import frc.robot.commands.FeederTestCommand;
+
 
 /**
  * This class is where the bulk of the robot should be declared. Since Command-based is a
@@ -18,13 +25,22 @@ import edu.wpi.first.wpilibj2.command.button.Trigger;
  */
 public class RobotContainer {
   // The robot's subsystems and commands are defined here...
-  
+  private Pivot pivot;
+  private Shooter shooter;
+  private Feeder feeder;
+  private final PivotTestCommand m_pivotTestCommand = new PivotTestCommand(pivot);
+  private final ShooterTestCommand m_shooterTestCommand = new ShooterTestCommand(shooter);
+  private final FeederTestCommand m_triggerTestCommand = new FeederTestCommand(feeder);
   // Replace with CommandPS4Controller or CommandJoystick if needed
   private final CommandXboxController m_driverController =
       new CommandXboxController(OperatorConstants.kDriverControllerPort);
 
   /** The container for the robot. Contains subsystems, OI devices, and commands. */
   public RobotContainer() {
+    CommandScheduler.getInstance().setDefaultCommand(pivot, m_pivotTestCommand);
+    CommandScheduler.getInstance().setDefaultCommand(shooter, m_shooterTestCommand);
+    //CommandScheduler.getInstance().setDefaultCommand(trigger, m_triggerTestCommand);
+
     // Configure the trigger bindings
     configureBindings();
   }
